+++ src/main/java/cn/nukkit/network/protocol/PlayerEnchantOptionsPacket.java
@@ -1,5 +1,7 @@
 package cn.nukkit.network.protocol;
 
+import cn.nukkit.api.DeprecationDetails;
+import cn.nukkit.api.PowerNukkitOnly;
 import cn.nukkit.api.Since;
 import it.unimi.dsi.fastutil.objects.ObjectArrayList;
 import lombok.ToString;
@@ -24,7 +26,7 @@ public class PlayerEnchantOptionsPacket extends DataPacket {
     public void decode() {
         int size = (int) this.getUnsignedVarInt();
         for (int i = 0; i < size; i++) {
-            int cost = this.getVarInt();
+            int minLevel = this.getVarInt();
             int slot = this.getInt();
 
             int eSize = (int) this.getUnsignedVarInt();
@@ -49,7 +51,7 @@ public class PlayerEnchantOptionsPacket extends DataPacket {
             }
             String enchantName = this.getString();
             int eNetId = (int) this.getUnsignedVarInt();
-            this.options.add(new EnchantOptionData(cost, slot, list1, list2, list3, enchantName, eNetId));
+            this.options.add(new EnchantOptionData(minLevel, slot, list1, list2, list3, enchantName, eNetId));
         }
 
     }
@@ -59,7 +61,7 @@ public class PlayerEnchantOptionsPacket extends DataPacket {
         this.reset();
         this.putUnsignedVarInt(this.options.size());
         for (EnchantOptionData option : this.options) {
-            this.putVarInt(option.getCost());
+            this.putVarInt(option.getMinLevel());
             this.putInt(option.getPrimarySlot());
             this.putUnsignedVarInt(option.getEnchants0().size());
             for (EnchantData data : option.getEnchants0()) {
@@ -85,13 +87,21 @@ public class PlayerEnchantOptionsPacket extends DataPacket {
     @Since("1.3.0.0-PN")
     @Value
     public class EnchantOptionData {
-        private final int cost;
+        private final int minLevel;
         private final int primarySlot;
         private final List<EnchantData> enchants0;
         private final List<EnchantData> enchants1;
         private final List<EnchantData> enchants2;
         private final String enchantName;
         private final int enchantNetId;
+
+        @PowerNukkitOnly("Does not exists in NukkitX anymore. Keeping for backward compatibility to allow smother version transition.")
+        @Deprecated @DeprecationDetails(since = "1.3.1.0-PN",
+                reason = "Renamed to minLevel by NukkitX", 
+                replaceWith = "getMinLevel()", toBeRemovedAt = "1.4.0.0-PN")
+        public int getCost() {
+            return getMinLevel();
+        }
     }
 
     @Since("1.3.0.0-PN")